+++ tests/end2endtest/utils/pytest_extensions.py
@@ -299,7 +299,7 @@ def wbem_connection(request, es_server):
         # Check that the server can be reached and authenticated with, by
         # issuing some quick operation. Operation failures are tolerated
         # (e.g. GetQualifier is not supported on SFCB), and the connection and
-        # authgentication errors can still be detected because they will be
+        # authentication errors can still be detected because they will be
         # detected before the operation fails.
         try:
             conn.GetQualifier(